% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grillage.R
\name{gr_create_shape}
\alias{gr_create_shape}
\title{Create shape}
\usage{
gr_create_shape(points, id, n = c(150, 150), code_insee = NULL)
}
\arguments{
\item{points}{an `sf` object with points}

\item{id}{the name of the column holding the IDs of the shapes}

\item{n}{a vector, number of cells used (rows and columns)}
}
\value{
an `sf` object with polygons
}
\description{
Create shape
}
