% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grillage.R
\name{gr_get_enveloppe}
\alias{gr_get_enveloppe}
\title{Utilitary function allowing to get the shape of a commune given its code Insee}
\usage{
gr_get_enveloppe(code_insee)
}
\description{
Utilitary function allowing to get the shape of a commune given its code Insee
}
