% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grillage.R
\name{gr_list_to_bv}
\alias{gr_list_to_bv}
\title{Get polling stations shapes from the voters register}
\usage{
gr_list_to_bv(
  path,
  code_insee,
  n = c(50, 50),
  confidence_level = 0.6,
  numero_bv = `numéro du bureau de vote`,
  numero_voie = NumeroVoie,
  libelle_voie = LibelleVoie,
  code_postal = `Code postal`,
  ville = VilleLocalite
)
}
\arguments{
\item{path}{path to the voters register (preferably in the REU format)}

\item{code_insee}{code INSEE of the commune}

\item{n}{a vector, number of cells used (rows and columns)}

\item{confidence_level}{threshold under which geolocalisation is not to be trusted}

\item{numero_bv}{column containing polling station code/number}

\item{numero_voie}{column containing the number in the street}

\item{libelle_voie}{column containing the name of the street}

\item{code_postal}{column containing the zip code}

\item{ville}{column containing the name of the city}
}
\value{
an sf object containing the (estimated) shape of the polling stations
}
\description{
Get polling stations shapes from the voters register
}
