% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grillage.R
\name{gr_lists_dpt}
\alias{gr_lists_dpt}
\title{Make polling stations polygons for a whole département}
\usage{
gr_lists_dpt(
  dpt,
  path = "/media/Data/listes électorales IDF 2020/",
  out_path = "."
)
}
\arguments{
\item{dpt}{the departement code}

\item{path}{the path to the folder where the voters registers are}
}
\value{
writes
}
\description{
Make polling stations polygons for a whole département
}
